dialog 위치 정리
