+++ app/src/main/java/womenproject/com/mybury/presentation/dialog/NetworkFailDialog.kt
@@ -1,0 +1,24 @@
+package womenproject.com.mybury.presentation.dialog
+
+import android.view.View
+import womenproject.com.mybury.presentation.base.BaseNormalDialogFragment
+
+class NetworkFailDialog : BaseNormalDialogFragment() {
+
+    init {
+        TITLE_MSG = "네트워크 통신 실패"
+        CONTENT_MSG = "네트워크가 불완전합니다. 다시 시도해주세요."
+        CANCEL_BUTTON_VISIBLE = false
+        GRADIENT_BUTTON_VISIBLE = true
+        CONFIRM_TEXT = "확인"
+        CANCEL_ABLE = false
+    }
+
+    override fun createOnClickConfirmListener(): View.OnClickListener {
+        return View.OnClickListener {
+            dismiss()
+            requireActivity().finish()
+        }
+    }
+
+}
